descartado/codigo_python/requirements.txt
mysql-connector-python
pymysql

# or

supabase-py